requirements.txt
google-genai>=1.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pillow>=11.0.0
